Remove "for" from start/end date labels
The "for" value pointed to an ID for an input that didn't exist, which
caused screenreaders to not read the value of the label when
specififying readthrough dates. Removing the
'for' tags makes the screenreader read the label as expected since the
label wraps the input, so we don't need a 'for'.