Fix migration for if db has multiple empty emails
If the database has multiple users with an empty email column, this
migration will fail because multiple empty strings break the unique
constraint.

A fresh database won't have this problem because it won't have any
legacy users with empty strings instead of NULL, but for existing
databases we need to convert the empty strings to NULL so they don't run
awry of the unique constraint.