URL-Encode password in broker_url
Needed to prevent connection errors if the password contains non-urlsafe characters.

Dockerfile change because all the services using it are mirroring ./ in /app now anyway.
Flower needs because it needs bookwyrm in order to read the celerywyrm config (because flower only allows overwriting broker_url from the command line and we want to do it in Python so we have access to a urlencode function).