Fix bw-dev resetdb, some tweaks to bw-dev
The main fix here is for resetdb, it previously was failing to drop the
db for me, because `web` was up and running and using the database. This
commit spins up db by itself first so it can drop and re-create the
database successfully, then brings up web to run the migrations.

While I was in here, I also updated it so that when running `bw-dev`
without any command it will also print the helptext, rather than just
exiting silently, got rid of the double-echo of the helptext, and added
runweb/rundb commands to run arbitrary commands via bw-dev.