Show deleted users as red in the user list
It can be hard to differentiate at a glance if a user is deleted or
suspended -- without this, you would have to read the deactivation
reason. By making deletions (moderator and self deletions) red, it's
clear at a glance if an account has been permanently deleted or just
temporarily suspended.