Show avatars in user admin view
I find this helpful as a rough hueristic for how engaged a user is -- if
they've never added an avatar, it's often the case that they never
really got further in to using the app than just registering a username.
Also, on one occassion, a user had an avatar that violated the instance
ToS, which would have been handy to see at a glance in this list.