Simplify identifier searches
This commit removes code that deduplicated search results for identifier
searches. If it was the case that multiple books have the same
identifier, in theory this would produce better search results, but in
practice this doesn't happen very much, is probably worth seeing when
it does, and worsens the performance of identifier search overall.