Unify docker-compose except for postgres
This adds two new files, docker-compose.prod.yml and
docker-compose.dev.yml, which contain any configuration specific to
prod or dev. They can be used as overrides along with the main
docker-compose file to create a full configuration.

This also removes all remaining ports from docker-compose, because
lists in override files are merged, not replaced, so we can add
things to lists but not remove them - so anything that's in a list
that we don't want to be universally present (such as ports) has to
exist only in the override files.

Other details: just moved the redis volume key down to be consistent
with main, moved the commented certbot command to the override file,
and defaulted to not using gunicorn, moving that into the prod file
as well.

As-is, with a plain docker-compose users will be able to start up a
stack and have it running, but not be able to use the app because it
won't be exposing any ports. At minimum they will need to add an
nginx (or web) port exposure.

We could move that one port into an environment variable, as an
in-between option.