Audiofic Archive and AO3 IDs
Introduces Audiofic Archive and AO3 IDs to books and authors.

Audiofic ID and AO3 ID will be shown in Ediitons view (people can see that there is an ID and therefore a link to read it)

Audiofic ID and AO3 ID will be whon on author page and will be linked

None of these IDs load any data at the moment, it is just to create the links and to show people the mere existence of these things.