oidc: Enable roles and update UI for OIDC-only login
If the OIDC server includes role information in the userinfo,
it will be checked for admin and moderator roles, which will be
applied to the user on each login.

If OIDC_ENABLED=true then the normal login forms will be
replaced with the (currently unstyled) "Login with OIDC"
link.

This also silences some of the debugging printouts.