temporarily fix signed get requests failing
Various issues have been raised regarding requests that seem to be dropped.
The real fix for this in #2717 - "Stop making HTTP requests in views"

This commit is a quick fix in some circumstances - particularly #2844 - signed GET requests failing in production.
It adds an extra thread to Gunicorn to make the default configuration multi-threaded.
This will now match the multi-threaded nature of the development environment, which is why we have not been able to replicate these errors in dev.