Shelter stable dates from USE_TZ
Some dates (publication dates, author dates) are meant as _literals_. What
the user inputs through a `SelectDateWidget` should be preserved as-is.
Django's otherwise-excelent support for timezones interferes with it (see

Until a better fate of these columns is determined (do we migrate them to a
DateField?), and as a stop-gap measure, we can start being faithful to the
data by storing them in the Eastern-most timezone.

This is particularly important because 1/1/YYYY is a common pattern in
publication dates, given #743.