Add exports_volume to docker-compose.yml
Exports should be written to a Docker volume instead of to the bind mount (= source directory). This
way they are shared between different containers even when they run on different machines.

After updating to the setup using the exports_volume, the following command can be used to copy
existing exports that have ended up in the source directory to the new Docker volume:

```sh
docker-compose cp exports web:/app
```