fix edition formatting and add alt link
Template blocks can't have logic or other blocks inside them.

This commit fixes the problem with the Edition template by removing the blocktrans block: original titles and descriptions shouldn't be translated, and this also makes the template simpler and avoids the logic-in-blocks problem.

Additionally if an edition is lacking a description, we fall back to the parent_work description, and line breaks have been added where needed.

Also adds a 'rel="alternate"' link to shelf pages to aid feed auto-discovery